refactor: PopularThemeDto -> PopularThemeResult로 변경
