+++ src/main/java/roomescape/repository/ThemeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
 import roomescape.domain.Theme;
-import roomescape.repository.dto.PopularThemeDto;
+import roomescape.repository.result.PopularThemeResult;
 
 import java.sql.PreparedStatement;
 import java.time.LocalDate;
@@ -54,7 +54,7 @@ public class ThemeRepository {
         return jdbcTemplate.update(sql, id);
     }
 
-    public List<PopularThemeDto> findPopular(LocalDate startDate, LocalDate endDate, int limit) {
+    public List<PopularThemeResult> findPopular(LocalDate startDate, LocalDate endDate, int limit) {
         String sql = """
                 SELECT
                     t.id,
@@ -83,8 +83,8 @@ public class ThemeRepository {
         return theme;
     };
 
-    private final RowMapper<PopularThemeDto> popularThemeRowMapper = (resultSet, rowNum) -> {
-        PopularThemeDto theme = new PopularThemeDto(
+    private final RowMapper<PopularThemeResult> popularThemeRowMapper = (resultSet, rowNum) -> {
+        PopularThemeResult theme = new PopularThemeResult(
                 resultSet.getLong("id"),
                 resultSet.getString("name"),
                 resultSet.getString("description"),
